docs: moved predicate methods (`simulateTransaction`) docs snippets
